+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased
 
-- [#138] Support form_post response mode
+- [#138] Support form_post response mode (This gem now requires Doorkeeper v5.5)
 
 ## v1.7.5 (2020-12-15)